src/main/java/com/gin/service/StudentService.java
package com.gin.service;

import com.gin.entity.Student;
import com.gin.repository.StudentRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class StudentService {

    private final StudentRepository studentRepository;

    public StudentService(StudentRepository studentRepository) {
        this.studentRepository = studentRepository;
    }

    public void initializeDatabase() {
        studentRepository.createTable();
    }

    public void saveStudent(Student student) {
        studentRepository.insertStudent(student);
    }

    public List<Student> getAllStudents() {
        return studentRepository.getAllStudents();
    }

    public Student getStudentById(Long id) {
        return studentRepository.getStudentById(id);
    }
}
